removed loss backward

+++ model1/dct_transformer.py
@@ -38,8 +38,8 @@ def transfom_to_multiple_images(image_to_transform):
     '''
     First resize into (64,64) and take DCT : it will convert into 
     '''
-    rgb_image_64 = resize(rgb_image_in_255, (64, 64))
-    rgb_image_32 = resize(rgb_image_in_255, (32, 32))
+    rgb_image_64 = resize(rgb_image_in_255, (64, 64)).astype(np.uint8)
+    rgb_image_32 = resize(rgb_image_in_255, (32, 32)).astype(np.uint8)
     # This resize has different output ???
     # rgb_image_64 = cv2.resize(rgb_image, (64, 64), interpolation=cv2.INTER_CUBIC)
     # rgb_image_32 = cv2.resize(rgb_image, (32, 32), interpolation=cv2.INTER_CUBIC)
